Test of running pre-commit in .ipynb files
The idea is that they get unstaged before commit.  But that didn't
seem to work.  So... I am now in the middle of an interactive rebase,
in an effort to sort out the mess.

I did "git reset HEAD~ **/*.ipynb" in order to unstage the notebook
files.  This set the state of the files back to the previous commit,
where they were not in the repo.